Code refactor and optimizing for PR
